Product Card Updates & product Modal
Resized images so that product cards would all be the same size.  Removed the truncated product descriptiuon from the product cards. Changed the card's add to cart button, and added a second "expander" button that opens a modal, to serve as a product page. The modal features a larger image and product description in addition to the price and product name displayed in the cards.