Shopping Cart Implementation
Refactored adding items to the cart. Now, if a product is already listed in the cart, the quantity of that product is incremented rather than creating a new line item.  Changed the styling of the jumbtron headings. Now, rather than black text with black borders, the text color is white and it is outlined in black, for a smoother look.

The quantities of items in the cart can now be changed from the view cart page. Items with quantities less than one remain in the array but are not displayed.

During the checkout process, the shipping address is autofilled from the users's profile, but the user has the opportunity to change shipping info.

The remaining issue with submitting an order at this time is that the insert  process is not resulting in a new row in the orders and order_items tables. Once that problem is addressed, the submit order script should be updated to unset the shopping cart array.