Final commit ( I think)
This should be the final product for the project. Mike's scripts have been refactored to check for the necessary userTypes before allowing access to critical admin features, changing some of the variables so that they grabbed data from the appropriate place, and changing some of the display scripts so that things like order histories displayed more usable information.

Additionally, order submission now correclty updates inventory for products.

There is no measuer in place for when inventory is below 0. This is currently treated as if the item is on backorder, but allows the 'purchase' to continue as normal